expose 'package' as standalone command. good for debugging/separation
