+++ package.json
@@ -21,15 +21,15 @@
   "engines": {
     "node": ">=8"
   },
-  "scripts": {
-    "test": "flt",
-    "clean": "rm -f index.js | rm -f sam.json | rm -f sam.yaml | rm -f sam-packaged.json | rm -f .eslintcache"
-  },
   "dependencies": {
     "js-yaml": "^3.11.0",
     "sade": "^1.4.1"
   },
   "devDependencies": {
     "flt": "^0.3.0"
+  },
+  "scripts": {
+    "test": "flt",
+    "clean": "rm -f index.js | rm -f sam.json | rm -f sam.yaml | rm -f sam-packaged.json | rm -f .eslintcache"
   }
 }
