+++ src/templates/lambda.js
